Rework setup and update Gem dependencies.
We now use a Gemset scheme for `rbenv` which gives us an explicit, repeatable process for bundling Jekyll that, assuming competent maintenance of the Jekyll Gem itself, should never fail to bundle and work properly. This also frees us from unnecessarily installing Gems into the system Gem repository that aren’t certain to be used elsewhere.

Note that this “explicit, repeatable process” *does not* depend on explicit version specifiers in the Gemfile; rather, we’re installing specific versions of those Gems into a Gemset which we point Bundler to when running `bundle install —local`.